feat(cli): add layered configuration merging and diag_json resolution
- Introduce `config_merge` module to handle CLI configuration discovery and merging
- Support layering of default, file, environment, and CLI overrides
- Preserve diagnostic JSON flag from merged layers for early error reporting
- Refactor CLI module to use new config merging utilities
- Add tests for theme resolution and output preferences

This improves configuration handling flexibility by cleanly merging multiple sources, enabling consistent CLI behavior and localization support.
