+++ src/cli/config_merge.rs
@@ -1,0 +1,177 @@
+//! CLI configuration discovery and merge helpers.
+//!
+//! This module keeps config-layer plumbing out of `cli::mod` so the public CLI
+//! surface stays focused on argument definitions and parsing.
+
+use clap::ArgMatches;
+use clap::parser::ValueSource;
+use ortho_config::declarative::LayerComposition;
+use ortho_config::figment::{Figment, providers::Env};
+use ortho_config::uncased::Uncased;
+use ortho_config::{
+    ConfigDiscovery, LocalizationArgs, MergeComposer, OrthoMergeExt, OrthoResult, sanitize_value,
+};
+use std::path::{Path, PathBuf};
+use std::sync::Arc;
+
+use crate::localization::{self, keys};
+
+use super::{CONFIG_ENV_VAR, Cli, ENV_PREFIX, validation_message};
+
+/// Return the default manifest filename when none is provided.
+pub(super) fn default_manifest_path() -> PathBuf {
+    PathBuf::from("Netsukefile")
+}
+
+/// Return the prefixed environment provider for CLI configuration.
+fn env_provider() -> Env {
+    Env::prefixed(ENV_PREFIX)
+}
+
+/// Build configuration discovery rooted in the optional working directory.
+fn config_discovery(directory: Option<&Path>) -> ConfigDiscovery {
+    let mut builder = ConfigDiscovery::builder("netsuke").env_var(CONFIG_ENV_VAR);
+    if let Some(dir) = directory {
+        builder = builder.clear_project_roots().add_project_root(dir);
+    }
+    builder.build()
+}
+
+/// Return `true` when no CLI overrides were supplied.
+///
+/// The merge pipeline treats an empty JSON object as "no overrides".
+fn is_empty_value(value: &serde_json::Value) -> bool {
+    matches!(value, serde_json::Value::Object(map) if map.is_empty())
+}
+
+fn diag_json_from_layer(value: &serde_json::Value) -> Option<bool> {
+    value
+        .as_object()
+        .and_then(|map| map.get("diag_json"))
+        .and_then(serde_json::Value::as_bool)
+}
+
+/// Resolve the effective diagnostic JSON preference from the raw config layers.
+///
+/// This is used before full config merging so startup and merge-time failures
+/// can still honour `diag_json` values sourced from config files or the
+/// environment.
+#[must_use]
+pub fn resolve_merged_diag_json(cli: &Cli, matches: &ArgMatches) -> bool {
+    let mut diag_json = Cli::default().diag_json;
+
+    let discovery = config_discovery(cli.directory.as_deref());
+    let file_layers = discovery.compose_layers();
+    for layer in file_layers.value {
+        let layer_value = layer.into_value();
+        if let Some(layer_diag_json) = diag_json_from_layer(&layer_value) {
+            diag_json = layer_diag_json;
+        }
+    }
+
+    let env_provider = env_provider()
+        .map(|key| Uncased::new(key.as_str().to_ascii_uppercase()))
+        .split("__");
+    if let Ok(value) = Figment::from(env_provider).extract::<serde_json::Value>()
+        && let Some(env_diag_json) = diag_json_from_layer(&value)
+    {
+        diag_json = env_diag_json;
+    }
+
+    if matches.value_source("diag_json") == Some(ValueSource::CommandLine) {
+        cli.diag_json
+    } else {
+        diag_json
+    }
+}
+
+fn cli_overrides_from_matches(cli: &Cli, matches: &ArgMatches) -> OrthoResult<serde_json::Value> {
+    let value = sanitize_value(cli)?;
+    let mut map = match value {
+        serde_json::Value::Object(map) => map,
+        other => {
+            let mut args = LocalizationArgs::default();
+            args.insert("value", format!("{other:?}").into());
+            let localizer = localization::localizer();
+            return Err(Arc::new(ortho_config::OrthoError::Validation {
+                key: String::from("cli"),
+                message: validation_message(
+                    localizer.as_ref(),
+                    keys::CLI_CONFIG_EXPECTED_OBJECT,
+                    Some(&args),
+                    &format!("expected parsed CLI values to serialize to an object, got {other:?}"),
+                ),
+            }));
+        }
+    };
+
+    map.remove("command");
+    for field in [
+        "file",
+        "verbose",
+        "fetch_default_deny",
+        "fetch_allow_scheme",
+        "fetch_allow_host",
+        "fetch_block_host",
+        "accessible",
+        "progress",
+        "no_emoji",
+        "theme",
+        "diag_json",
+    ] {
+        if matches.value_source(field) != Some(ValueSource::CommandLine) {
+            map.remove(field);
+        }
+    }
+
+    Ok(serde_json::Value::Object(map))
+}
+
+/// Merge configuration layers over the parsed CLI values.
+///
+/// # Errors
+///
+/// Returns an [`ortho_config::OrthoError`] if layer composition or merging
+/// fails.
+pub fn merge_with_config(cli: &Cli, matches: &ArgMatches) -> OrthoResult<Cli> {
+    let command = cli.command.clone();
+    let mut errors = Vec::new();
+    let mut composer = MergeComposer::with_capacity(4);
+
+    match sanitize_value(&Cli::default()) {
+        Ok(value) => composer.push_defaults(value),
+        Err(err) => errors.push(err),
+    }
+
+    let discovery = config_discovery(cli.directory.as_deref());
+    let mut file_layers = discovery.compose_layers();
+    errors.append(&mut file_layers.required_errors);
+    if file_layers.value.is_empty() {
+        errors.append(&mut file_layers.optional_errors);
+    }
+    for layer in file_layers.value {
+        composer.push_layer(layer);
+    }
+
+    let env_provider = env_provider()
+        .map(|key| Uncased::new(key.as_str().to_ascii_uppercase()))
+        .split("__");
+    match Figment::from(env_provider)
+        .extract::<serde_json::Value>()
+        .into_ortho_merge()
+    {
+        Ok(value) => composer.push_environment(value),
+        Err(err) => errors.push(err),
+    }
+
+    match cli_overrides_from_matches(cli, matches) {
+        Ok(value) if !is_empty_value(&value) => composer.push_cli(value),
+        Ok(_) => {}
+        Err(err) => errors.push(err),
+    }
+
+    let composition = LayerComposition::new(composer.layers(), errors);
+    let mut merged = composition.into_merge_result(Cli::merge_from_layers)?;
+    merged.command = command;
+    Ok(merged)
+}
